refactor: move register component to components/register-page
