+++ src/app/(sign)/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegisterBox } from '@/components/signUpIn/RegisterBox';
+import { RegisterBox } from '@/components/register-page/RegisterBox';
 
 export default function Page() {
   return (
